Updated default path for camellia library (again)
I installed camcryp via pip on Linux and also via easy_install on a Mac. This resulted in some very strange string handling of passwords when passed to the camellia library which led to incorrect Snort signatures being created. Replacing the repository version of camcrypt with a locally built version cloned from git fixed the issue.